+++ spec/dummy/app/models/store/manufacturer.rb
@@ -1,0 +1,5 @@
+module Store
+  class Manufacturer < ActiveRecord::Base
+    has_many :cars
+  end
+end