+++ spec/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170827225107) do
+ActiveRecord::Schema.define(version: 20170904155843) do
 
   create_table "active_admin_comments", force: true do |t|
     t.string   "namespace"
@@ -45,6 +45,12 @@ ActiveRecord::Schema.define(version: 20170827225107) do
 
   add_index "admin_users", ["email"], name: "index_admin_users_on_email", unique: true
   add_index "admin_users", ["reset_password_token"], name: "index_admin_users_on_reset_password_token", unique: true
+
+  create_table "cars", force: true do |t|
+    t.string  "name"
+    t.integer "year"
+    t.integer "manufacturer_id"
+  end
 
   create_table "categories", force: true do |t|
     t.string   "name"
@@ -111,6 +117,10 @@ ActiveRecord::Schema.define(version: 20170827225107) do
     t.text     "description"
     t.datetime "created_at"
     t.datetime "updated_at"
+  end
+
+  create_table "manufacturers", force: true do |t|
+    t.string "name"
   end
 
   create_table "regions", force: true do |t|